Refactor telemetry code to improve re-compilation times
